# swagger_server/models/rental.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Rental(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, movie_id: str=None, user_id: str=None, rental_date: datetime=None, return_date: datetime=None):  # noqa: E501
        """Rental - a model defined in Swagger

        :param id: The id of this Rental.  # noqa: E501
        :type id: str
        :param movie_id: The movie_id of this Rental.  # noqa: E501
        :type movie_id: str
        :param user_id: The user_id of this Rental.  # noqa: E501
        :type user_id: str
        :param rental_date: The rental_date of this Rental.  # noqa: E501
        :type rental_date: datetime
        :param return_date: The return_date of this Rental.  # noqa: E501
        :type return_date: datetime
        """
        self.swagger_types = {
            'id': str,
            'movie_id': str,
            'user_id': str,
            'rental_date': datetime,
            'return_date': datetime
        }

        self.attribute_map = {
            'id': 'id',
            'movie_id': 'movieId',
            'user_id': 'userId',
            'rental_date': 'rentalDate',
            'return_date': 'returnDate'
        }
        self._id = id
        self._movie_id = movie_id
        self._user_id = user_id
        self._rental_date = rental_date
        self._return_date = return_date

    @classmethod
    def from_dict(cls, dikt) -> 'Rental':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Rental of this Rental.  # noqa: E501
        :rtype: Rental
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Rental.

        Уникальный идентификатор аренды  # noqa: E501

        :return: The id of this Rental.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Rental.

        Уникальный идентификатор аренды  # noqa: E501

        :param id: The id of this Rental.
        :type id: str
        """

        self._id = id

    @property
    def movie_id(self) -> str:
        """Gets the movie_id of this Rental.

        ID арендованного фильма  # noqa: E501

        :return: The movie_id of this Rental.
        :rtype: str
        """
        return self._movie_id

    @movie_id.setter
    def movie_id(self, movie_id: str):
        """Sets the movie_id of this Rental.

        ID арендованного фильма  # noqa: E501

        :param movie_id: The movie_id of this Rental.
        :type movie_id: str
        """

        self._movie_id = movie_id

    @property
    def user_id(self) -> str:
        """Gets the user_id of this Rental.

        ID пользователя, арендующего фильм  # noqa: E501

        :return: The user_id of this Rental.
        :rtype: str
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: str):
        """Sets the user_id of this Rental.

        ID пользователя, арендующего фильм  # noqa: E501

        :param user_id: The user_id of this Rental.
        :type user_id: str
        """

        self._user_id = user_id

    @property
    def rental_date(self) -> datetime:
        """Gets the rental_date of this Rental.

        Дата и время начала аренды  # noqa: E501

        :return: The rental_date of this Rental.
        :rtype: datetime
        """
        return self._rental_date

    @rental_date.setter
    def rental_date(self, rental_date: datetime):
        """Sets the rental_date of this Rental.

        Дата и время начала аренды  # noqa: E501

        :param rental_date: The rental_date of this Rental.
        :type rental_date: datetime
        """

        self._rental_date = rental_date

    @property
    def return_date(self) -> datetime:
        """Gets the return_date of this Rental.

        Дата и время окончания аренды  # noqa: E501

        :return: The return_date of this Rental.
        :rtype: datetime
        """
        return self._return_date

    @return_date.setter
    def return_date(self, return_date: datetime):
        """Sets the return_date of this Rental.

        Дата и время окончания аренды  # noqa: E501

        :param return_date: The return_date of this Rental.
        :type return_date: datetime
        """

        self._return_date = return_date